scripts/run_streamlit.py
#!/usr/bin/env python3
import os
import sys
from pathlib import Path
from dotenv import load_dotenv

# Add project root to Python path
project_root = Path(__file__).parent.parent
sys.path.append(str(project_root))

# Load environment variables
load_dotenv(project_root / '.env')

# Import streamlit app
from src.streamlit_app import main

if __name__ == "__main__":
    main() 