requirements.txt
# Database
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9

# API
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9

# Data Processing
pandas==2.2.0
numpy==1.26.4

# Caching
redis==5.0.1
hiredis==2.3.2

# Validation & Serialization
pydantic==2.6.1
python-dateutil==2.8.2

# Environment & Configuration
python-dotenv==1.0.1
pyyaml==6.0.1

# Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2

# Monitoring & Logging
prometheus-client==0.19.0
structlog==24.1.0

# Testing
pytest==8.0.0
pytest-cov==4.1.0
pytest-asyncio==0.23.5

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0 